Finii

+++ ft_char_x.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_char_x.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yel-mens <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: yel-mens <yel-mens@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/21 15:13:58 by yel-mens          #+#    #+#             */
-/*   Updated: 2024/10/21 19:10:34 by yel-mens         ###   ########.fr       */
+/*   Updated: 2024/10/22 17:59:54 by yel-mens         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-static void	ft_printptr(unsigned long int ptr, int *count)
+static void	ft_printptr(unsigned long ptr, int *count)
 {
 	if (ptr < 16)
 	{
@@ -28,8 +28,8 @@ static void	ft_printptr(unsigned long int ptr, int *count)
 
 void	ft_char_x(va_list args, int *count)
 {
-	unsigned long int	ptr;
+	unsigned int	ptr;
 
-	ptr = va_arg(args, unsigned long int);
-	ft_printptr(ptr, count);
+	ptr = va_arg(args, unsigned int);
+	ft_printptr((unsigned long)ptr, count);
 }
